Throw a new .distributionDoesNotSupportArchitecture error when attempting to create a Swift SDK with rhel for armv7

+++ Sources/SwiftSDKGenerator/SwiftSDKRecipes/LinuxRecipe.swift
@@ -211,6 +211,12 @@ public struct LinuxRecipe: SwiftSDKRecipe {
     engine: QueryEngine,
     httpClient client: some HTTPClientProtocol
   ) async throws -> SwiftSDKProduct {
+    if self.linuxDistribution.name == .rhel && self.mainTargetTriple.archName == "armv7" {
+      throw GeneratorError.distributionDoesNotSupportArchitecture(
+        self.linuxDistribution, targetArchName: self.mainTargetTriple.archName
+      )
+    }
+  
     let sdkDirPath = self.sdkDirPath(paths: generator.pathsConfiguration)
     if !generator.isIncremental {
       try await generator.removeRecursively(at: sdkDirPath)
